refactor: move classes about ModificationsContainer to another file
